moved variables into residence.py
